refactor(daemon): move server to separate module
